eolo-crypto/requirements.txt
# EOLO Crypto — Dependencies
# Install: pip install -r requirements.txt

# HTTP (Binance REST + helpers)
requests>=2.31.0

# WebSocket (Binance combined streams)
websockets>=12.0

# Logging
loguru>=0.7.0

# Data / Math (estrategias Eolo v1 reutilizadas)
numpy>=1.26.0
pandas>=2.0.0

# Timezone (usado por estrategias de Bot/ que chequean market hours ET)
pytz>=2024.1
tzdata>=2024.1

# Claude API (bot #14 decision engine)
anthropic>=0.30.0

# Google Cloud (secretos + firestore para dashboard)
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.20.0

# Cloud Run entry point
flask>=3.0.0
gunicorn>=22.0.0
